fix(prd-api): 说话人盖戳断言改用既有已验证的合成音频，删掉自造的第二份
CI 只挂了这一条：我为它另写了一份 BuildWav（谐波、幅度、取整都和既有的不同），
声学聚类结果因此不一样。判据不该建在自己新造、没人验证过的地基上——
断言移进 LocalSpeakerDiarizerTests，复用那段已被证明能分出两个人的音频，
本文件只留纯文本可判定的部分。
